requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
gunicorn==22.0.0

langchain>=0.3.0,<0.4
langchain-openai>=0.3.0,<0.4

azure-identity>=1.17.0,<2
azure-ai-projects>=2.0.0
azure-communication-messages>=1.1.0
azure-communication-callautomation>=1.3.0
openai>=1.0.0